refactor: add prompts for volumes:create
also includes few help message fixes for droplets:create
